requirements.txt
face-recognition==1.3.0
imutils==0.5.3
Keras==2.4.3
matplotlib==3.3.3
opencv-python==4.4.0.46
pandas==1.2.0
scikit-learn==0.24.0
tensorflow
tensorflow-gpu
scipy==1.6.0
pillow
requests==2.24.0
h5py==2.10.0
